Aggregate software updates fetching in the SoftwareUpdates context
