+++ lib/trento_web/controllers/v1/suse_manager_controller.ex
@@ -2,9 +2,6 @@ defmodule TrentoWeb.V1.SUSEManagerController do
   use TrentoWeb, :controller
   use OpenApiSpex.ControllerSpecs
 
-  alias Trento.Hosts
-  alias Trento.Hosts.Projections.HostReadModel
-
   alias Trento.SoftwareUpdates
 
   alias TrentoWeb.OpenApi.V1.Schema
@@ -35,20 +32,13 @@ defmodule TrentoWeb.V1.SUSEManagerController do
 
   @spec software_updates(Plug.Conn.t(), any) :: Plug.Conn.t()
   def software_updates(conn, %{id: host_id}) do
-    with {:ok, fqdn} <- get_host_fqdn(host_id),
-         {:ok, system_id} <- SoftwareUpdates.Discovery.get_system_id(fqdn),
-         {:ok, relevant_patches} <- SoftwareUpdates.Discovery.get_relevant_patches(system_id),
-         {:ok, upgradable_packages} <-
-           SoftwareUpdates.Discovery.get_upgradable_packages(system_id) do
+    with {:ok,
+          %{
+            relevant_patches: relevant_patches,
+            upgradable_packages: upgradable_packages
+          }} <-
+           SoftwareUpdates.get_software_updates(host_id) do
       render(conn, %{relevant_patches: relevant_patches, upgradable_packages: upgradable_packages})
-    end
-  end
-
-  defp get_host_fqdn(host_id) do
-    case Hosts.get_host_by_id(host_id) do
-      nil -> {:error, :not_found}
-      %HostReadModel{fully_qualified_domain_name: nil} -> {:error, :fqdn_not_found}
-      %HostReadModel{fully_qualified_domain_name: fqdn} -> {:ok, fqdn}
     end
   end
 end
